allow to pin, archive, etc. when searching in the chatlist
