+++ res/layout/fragment_search.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -17,4 +18,14 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
+    <org.thoughtcrime.securesms.components.registration.PulsingFloatingActionButton
+            android:id="@+id/fab"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom|end"
+            android:layout_margin="16dp"
+            app:backgroundTint="?attr/fab_color"
+            android:src="@drawable/ic_add_white_24dp"
+            android:focusable="true"
+            android:contentDescription="@string/menu_new_chat"/>
 </FrameLayout>
