+++ res/menu/conversation_list.xml
@@ -2,6 +2,16 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto">
 
+    <item android:title="@string/pin_chat"
+          android:id="@+id/menu_pin_selected"
+          android:icon="@drawable/ic_pin_white"
+          app:showAsAction="always"/>
+
+    <item android:title="@string/menu_archive_chat"
+          android:id="@+id/menu_archive_selected"
+          android:icon="@drawable/ic_archive_white_24dp"
+          app:showAsAction="always"/>
+
     <item android:title="@string/menu_delete_chat"
           android:id="@+id/menu_delete_selected"
           android:icon="?menu_trash_icon"
