--- res/menu/conversation_list_batch_archive.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item android:title="@string/menu_archive_chat"
-          android:id="@+id/menu_archive_selected"
-          android:icon="@drawable/ic_archive_white_24dp"
-          app:showAsAction="always"/>
-
-</menu>
