--- res/menu/conversation_list_batch_pin.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item android:title="@string/pin_chat"
-          android:id="@+id/menu_pin_selected"
-          android:icon="@drawable/ic_pin_white"
-          app:showAsAction="always"/>
-
-</menu>
